Fix VFS issues: type safety, duplicate checks, and debug output
- Fix serial_putc type mismatch (i8 -> u8) across all VFS modules
- Add duplicate name checks in RamFS mkdir and open operations
- Add comprehensive debug output for mkdir and file creation paths
- Improve code robustness and debugging capabilities

+++ src/fs/vfs/vfs.rs
@@ -5,13 +5,13 @@ use core::sync::atomic::{AtomicU32, Ordering};
 use super::types::*;
 
 extern "C" {
-    fn serial_putc(port: u16, c: i8);
+    fn serial_putc(port: u16, c: u8);
 }
 
 fn log(s: &str) {
     unsafe {
         for c in s.bytes() {
-            serial_putc(0x3F8, c as i8);
+            serial_putc(0x3F8, c);
         }
     }
 }
